Reverted skipBytes/readBytes block-based logic
Having the skipBytes and readBytes methods that are aware of the header
blocks was redundant, given the case logic that takes care of skipping
to the appropriate locations for each data block.

Although it would be nice to standardize the logic for Csarseven and
FIFO, the difference between block per pixel and block per channel is
huge.